+++ test/lib/tags/callback_test.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:xwidget/xwidget.dart';
+
+import '../../fixtures/src/inflaters.g.dart';
+import '../testing_utils.dart';
+
+main() {
+  final assetBundle = TestAssetBundle([
+    "test/fixtures/resources"
+  ]);
+
+  setUpAll(() async {
+    TestWidgetsFlutterBinding.ensureInitialized();
+    await Resources.instance.loadResources("test/fixtures/resources", assetBundle);
+    registerXWidgetInflaters();
+  });
+
+  testWidgets('Assert callback action on TextButton fires', (tester) async {
+    const xml = '''
+      <MaterialApp>
+        <Scaffold for="home">
+          <Column for="body">
+            <TextButton>
+              <callback for="onPressed" action="doSomething('Hello world!')" returnVar="result"/>                 
+              <Text data="Press Me"/>
+            </TextButton>
+          </Column>
+        </Scaffold>
+      </MaterialApp>
+    ''';
+
+    final dependencies = Dependencies({
+      "doSomething": (msg)  => "Your message: $msg"
+    });
+
+    final testApp = XWidget.inflateFromXml(xml: xml, dependencies: dependencies);
+    await tester.pumpWidget(testApp);
+    await tester.tap(find.byType(TextButton));
+    expect(dependencies["result"], "Your message: Hello world!");
+  });
+}